Address PR #256 review: regular/ directory + a --regex test run
- Move the Regular constraint and its regex helper into a
  gcs/constraints/regular/ directory, mirroring the sort/ layout: regular.cc /
  regular.hh / regular_test.cc, plus the regex->NFA helper as regex.cc /
  regex.hh / regex_test.cc (dropping the now-redundant regular_ prefix). The
  public <gcs/constraints/regular.hh> becomes a thin forwarding header, so
  existing includers are unaffected. The regex_regex_test target is renamed
  regex_test.
- Add a regex-regex ctest that runs the example with --regex, exercising the
  regular-expression-string constructor end to end (it was previously only
  built, never run). Both regex runs share the regex.{opb,pbp} basename, so
  they take a RESOURCE_LOCK to avoid racing under parallel ctest (crystal_maze
  pattern).
